% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{embark_official}
\alias{embark_official}
\title{Embark official color palette}
\format{
An object of class \code{list} of length 19.
}
\source{
https://embark-theme.github.io/
}
\usage{
embark_official
}
\description{
These are the official colors as listed on
the \href{https://embark-theme.github.io/}{Embark Webpage}
in June 2023.
}
\keyword{datasets}
