% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{embark_palette}
\alias{embark_palette}
\title{Palette shortcuts to use with \code{rsthemes}}
\format{
An object of class \code{list} of length 38.
}
\usage{
embark_palette
}
\description{
Palette shortcuts to use with \code{rsthemes}
}
\keyword{datasets}
