% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{embark_rstheme}
\alias{embark_rstheme}
\title{Create Embark .rstheme file}
\usage{
embark_rstheme(install = TRUE)
}
\arguments{
\item{install}{A Boolean - generate and install/apply the .rstheme file
(TRUE), or generate the .scss file without installing/applying}
}
\description{
Create Embark .rstheme file
}
